+++ relay-server/src/processing/spans/process.rs
@@ -1,15 +1,17 @@
 use std::collections::BTreeMap;
 use std::time::Duration;
 
-use relay_event_normalization::{GeoIpLookup, ModelMetadata, RequiredMode, SchemaProcessor, eap};
+use relay_event_normalization::eap::ClientUserAgentInfo;
+use relay_event_normalization::{GeoIpLookup, RequiredMode, SchemaProcessor, eap};
 use relay_event_schema::processor::{ProcessingState, ValueType, process_value};
 use relay_event_schema::protocol::{Span, SpanId, SpanV2};
 use relay_protocol::Annotated;
 
 use crate::envelope::{ContainerItems, EnvelopeHeaders, Item, ItemContainer, ParentId, WithHeader};
-use crate::managed::Managed;
+use crate::managed::{Managed, RecordKeeper, Rejected};
 use crate::processing::spans::{
     self, Error, ExpandedAttachment, ExpandedSpan, ExpandedSpans, Result, SerializedSpans,
+    Settings, SpanItems,
 };
 use crate::processing::{Context, trace_attachments, utils};
 use crate::services::outcome::DiscardReason;
@@ -17,35 +19,29 @@ use crate::services::outcome::DiscardReason;
 /// Parses all serialized spans.
 ///
 /// Individual, invalid spans are discarded.
-pub fn expand(spans: Managed<SerializedSpans>) -> Managed<ExpandedSpans> {
-    spans.map(|spans, records| {
+pub fn expand(spans: Managed<SerializedSpans>) -> Result<Managed<ExpandedSpans>, Rejected<Error>> {
+    spans.try_map(|spans, records| {
         let SerializedSpans {
             headers,
-            spans,
-            legacy,
-            integrations,
+            items,
+            invalid,
             attachments,
         } = spans;
 
-        let mut all_spans = Vec::new();
+        debug_assert!(
+            invalid.is_empty(),
+            "invalid items should already be rejected"
+        );
 
-        for item in &spans {
-            let expanded = expand_span_container(item);
-            let expanded = records.or_default(expanded, item);
-            all_spans.extend(expanded);
-        }
-
-        for item in &legacy {
-            match expand_legacy_span(item) {
-                Ok(span) => all_spans.push(span),
-                Err(err) => drop(records.reject_err(err, item)),
-            }
-        }
-
-        spans::integrations::expand_into(&mut all_spans, records, integrations);
+        let (settings, spans) = match items {
+            SpanItems::Container(item) => expand_span_container(&item)?,
+            SpanItems::Legacy(items) => expand_legacy_spans(items, records),
+            SpanItems::Integration(item) => spans::integrations::expand(records, &[item]),
+            SpanItems::None => (Default::default(), Vec::new()),
+        };
 
         let mut span_id_mapping: BTreeMap<_, _> = BTreeMap::new();
-        for span in all_spans {
+        for span in spans {
             if let Some(id) = span.value().and_then(|span| span.span_id.value().copied()) {
                 // Although span_ids should be unique it could be that they collied in which case we
                 // want to drop one of the offending spans.
@@ -82,25 +78,67 @@ pub fn expand(spans: Managed<SerializedSpans>) -> Managed<ExpandedSpans> {
             }
         }
 
-        ExpandedSpans {
+        Ok::<_, Error>(ExpandedSpans {
             headers,
+            settings,
             spans: span_id_mapping.into_values().collect(),
             server_sample_rate: None,
             stand_alone_attachments,
             category: spans::TotalAndIndexed,
-        }
+        })
     })
 }
 
-fn expand_span_container(item: &Item) -> Result<ContainerItems<SpanV2>> {
-    let spans = ItemContainer::parse(item)
+fn expand_span_container(item: &Item) -> Result<(Settings, ContainerItems<SpanV2>)> {
+    let (metadata, spans) = ItemContainer::<SpanV2>::parse(item)
         .map_err(|err| {
             relay_log::debug!("failed to parse span container: {err}");
             Error::Invalid(DiscardReason::InvalidJson)
         })?
-        .into_items();
+        .into_parts();
 
-    Ok(spans)
+    relay_log::trace!("span container metadata: {metadata:?}");
+    let settings = metadata
+        .map(|metadata| {
+            let is = metadata.ingest_settings.as_ref();
+
+            match metadata.version {
+                None => Settings::default(),
+                // Technically invalid.
+                Some(0 | 1) => Settings::default(),
+                Some(2) => Settings {
+                    infer_ip: is
+                        .and_then(|is| is.infer_ip)
+                        .is_some_and(|infer| infer.is_auto()),
+                    infer_user_agent: is
+                        .and_then(|is| is.infer_user_agent)
+                        .is_some_and(|infer| infer.is_auto()),
+                },
+                // Unsupported, fall back to the safe default.
+                Some(_) => Default::default(),
+            }
+        })
+        .unwrap_or_default();
+
+    Ok((settings, spans))
+}
+
+fn expand_legacy_spans(
+    items: Vec<Item>,
+    records: &mut RecordKeeper<'_>,
+) -> (Settings, ContainerItems<SpanV2>) {
+    let spans = items
+        .into_iter()
+        .filter_map(|item| match expand_legacy_span(&item) {
+            Ok(span) => Some(span),
+            Err(err) => {
+                records.reject_err(err, item);
+                None
+            }
+        })
+        .collect();
+
+    (Settings::default(), spans)
 }
 
 fn expand_legacy_span(item: &Item) -> Result<WithHeader<SpanV2>> {
@@ -134,18 +172,12 @@ fn parse_and_validate_span_attachment(item: &Item) -> Result<(Option<SpanId>, Ex
 
 /// Normalizes individual spans.
 pub fn normalize(spans: &mut Managed<ExpandedSpans>, geo_lookup: &GeoIpLookup, ctx: Context<'_>) {
-    let model_metadata = ctx.global_config.model_metadata();
+    let settings = spans.settings;
+
     spans.retain_with_context(
         |spans| (&mut spans.spans, &spans.headers),
         |span, headers, _| {
-            normalize_span(
-                &mut span.span,
-                headers,
-                geo_lookup,
-                model_metadata.as_ref(),
-                ctx,
-            )
-            .inspect_err(|err| {
+            normalize_span(&mut span.span, settings, headers, geo_lookup, ctx).inspect_err(|err| {
                 relay_log::debug!("failed to normalize span: {err}");
             })
         },
@@ -154,9 +186,9 @@ pub fn normalize(spans: &mut Managed<ExpandedSpans>, geo_lookup: &GeoIpLookup, c
 
 fn normalize_span(
     span: &mut Annotated<SpanV2>,
+    settings: Settings,
     headers: &EnvelopeHeaders,
     geo_lookup: &GeoIpLookup,
-    model_metadata: Option<&ModelMetadata>,
     ctx: Context<'_>,
 ) -> Result<()> {
     let meta = headers.meta();
@@ -170,27 +202,33 @@ fn normalize_span(
         let dsc = headers.dsc();
         let duration = span_duration(span);
         let allowed_hosts = ctx.global_config.options.http_span_allowed_hosts.as_slice();
+        let model_metdata = ctx.global_config.ai_model_metadata();
+        let client_ua_info = settings.infer_user_agent.then(|| ClientUserAgentInfo {
+            user_agent: meta.user_agent(),
+            hints: meta.client_hints(),
+        });
 
         validate_timestamps(span)?;
 
+        eap::normalize_attribute_types(&mut span.attributes);
+        eap::normalize_attribute_names(&mut span.attributes);
         // normalize_sentry_op must be called before normalize_span_category
         // because category derivation depends on having the sentry.op attribute
         // available.
         eap::normalize_sentry_op(&mut span.attributes);
         eap::normalize_span_category(&mut span.attributes);
-        eap::normalize_attribute_types(&mut span.attributes);
-        eap::normalize_attribute_names(&mut span.attributes);
         eap::normalize_received(&mut span.attributes, meta.received_at());
         eap::normalize_client_address(&mut span.attributes, meta.client_addr());
-        eap::normalize_user_agent(&mut span.attributes, meta.user_agent(), meta.client_hints());
-        eap::normalize_user_geo(&mut span.attributes, || {
-            meta.client_addr().and_then(|ip| geo_lookup.lookup(ip))
-        });
+        if settings.infer_ip {
+            eap::normalize_inject_client_address(&mut span.attributes, meta.client_addr());
+        }
+        eap::normalize_user_agent(&mut span.attributes, client_ua_info);
+        eap::normalize_user_geo(&mut span.attributes, |ip| geo_lookup.lookup(ip));
         if matches!(span.is_segment.value(), Some(true)) {
             eap::normalize_dsc(&mut span.attributes, dsc);
         }
         if ctx.is_processing() {
-            eap::normalize_ai(&mut span.attributes, duration, model_metadata);
+            eap::normalize_ai(&mut span.attributes, duration, model_metdata);
         }
         eap::normalize_attribute_values(&mut span.attributes, allowed_hosts);
         eap::write_legacy_attributes(&mut span.attributes);
@@ -295,18 +333,8 @@ fn span_duration(span: &SpanV2) -> Option<Duration> {
 #[cfg(test)]
 mod tests {
     use chrono::DateTime;
-    #[allow(
-        deprecated,
-        reason = "This deprecated attribute is intentionally used in some legacy tests."
-    )]
-    use relay_conventions::DB__SYSTEM;
-    use relay_conventions::{
-        APP__VITALS__START__TYPE, APP__VITALS__START__VALUE, APP__VITALS__TTFD__VALUE,
-        DB__QUERY__TEXT, DB__SYSTEM__NAME, DEVICE__CLASS, DEVICE__FAMILY, DEVICE__MODEL,
-        HTTP__REQUEST__METHOD, SENTRY__ACTION, SENTRY__CATEGORY, SENTRY__DESCRIPTION,
-        SENTRY__DOMAIN, SENTRY__MAIN_THREAD, SENTRY__MOBILE, SENTRY__NORMALIZED_DESCRIPTION,
-        SENTRY__OP, SENTRY__SDK__NAME, THREAD__NAME, URL__FULL,
-    };
+
+    use relay_conventions::*;
     use relay_event_schema::protocol::{Attributes, EventId, SpanKind};
     use relay_pii::PiiConfig;
     use relay_protocol::SerializableAnnotated;
@@ -847,10 +875,18 @@ mod tests {
     ) {
         let attrs = span.value().unwrap().attributes.value().unwrap();
         string_attributes.iter().for_each(|(key, value)| {
-            assert_eq!(attrs.get_value(*key).and_then(|v| v.as_str()), Some(*value),)
+            assert_eq!(
+                attrs.get_value(*key).and_then(|v| v.as_str()),
+                Some(*value),
+                "attribute mismatch for {key}"
+            )
         });
         float_attributes.iter().for_each(|(key, value)| {
-            assert_eq!(attrs.get_value(*key).and_then(|v| v.as_f64()), Some(*value),)
+            assert_eq!(
+                attrs.get_value(*key).and_then(|v| v.as_f64()),
+                Some(*value),
+                "attribute mismatch for {key}"
+            )
         });
     }
 
@@ -865,7 +901,7 @@ mod tests {
             &[],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -895,7 +931,7 @@ mod tests {
             &[],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -925,7 +961,7 @@ mod tests {
             &[("http.response.status_code", 502.)],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -956,7 +992,7 @@ mod tests {
             &[("http.response.status_code", 502.)],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -989,7 +1025,7 @@ mod tests {
             ],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -1021,7 +1057,7 @@ mod tests {
             &[("http.response.status_code", 502.)],
         );
 
-        normalize_span(&mut span, &headers, &geo_lookup, None, ctx).unwrap();
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
 
         assert_attributes_contains(
             &span,
@@ -1033,6 +1069,83 @@ mod tests {
                 (SENTRY__DOMAIN, "*.example.com"),
             ],
             &[("sentry.status_code", 502.)],
+        );
+    }
+    #[test]
+    #[allow(
+        deprecated,
+        reason = "This test deliberately checks deprecated attributes"
+    )]
+    fn test_op_from_deprecated_db_system() {
+        let (mut span, headers, geo_lookup, ctx) = prepare_normalize_span_params(
+            &[
+                (DB__SYSTEM, "postgresql"),
+                (DB__QUERY__TEXT, "select * from users where id = 1"),
+            ],
+            &[],
+        );
+
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
+
+        assert_attributes_contains(
+            &span,
+            &[
+                (SENTRY__OP, "db"),
+                (SENTRY__DESCRIPTION, "select * from users where id = 1"),
+                (
+                    SENTRY__NORMALIZED_DESCRIPTION,
+                    "SELECT * FROM users WHERE id = %s",
+                ),
+                (SENTRY__CATEGORY, "db"),
+                (DB__SYSTEM__NAME, "postgresql"),
+                (SENTRY__ACTION, "SELECT"),
+                (SENTRY__DOMAIN, ",users,"),
+            ],
+            &[],
+        );
+    }
+
+    #[test]
+    #[allow(
+        deprecated,
+        reason = "This test deliberately checks deprecated attributes"
+    )]
+    fn test_op_from_deprecated_http_method() {
+        let (mut span, headers, geo_lookup, ctx) =
+            prepare_normalize_span_params(&[(HTTP__METHOD, "GET"), (SENTRY__KIND, "server")], &[]);
+
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
+
+        assert_attributes_contains(
+            &span,
+            &[
+                (SENTRY__OP, "http.server"),
+                (SENTRY__CATEGORY, "http"),
+                (HTTP__REQUEST__METHOD, "GET"),
+                (SENTRY__KIND, "server"),
+            ],
+            &[],
+        );
+    }
+
+    #[test]
+    #[allow(
+        deprecated,
+        reason = "This test deliberately checks deprecated attributes"
+    )]
+    fn test_op_from_deprecated_gen_ai_system() {
+        let (mut span, headers, geo_lookup, ctx) =
+            prepare_normalize_span_params(&[(GEN_AI__SYSTEM, "some system")], &[]);
+
+        normalize_span(&mut span, Default::default(), &headers, &geo_lookup, ctx).unwrap();
+
+        assert_attributes_contains(
+            &span,
+            &[
+                (SENTRY__OP, "gen_ai"),
+                (GEN_AI__PROVIDER__NAME, "some system"),
+            ],
+            &[],
         );
     }
 }